moved mkdir from util to fs. its just like original fs.mkdir but doesnt throw an error if the directory already exists
