+++ lib/fs.js
@@ -1,6 +1,5 @@
 import _fs from 'fs';
 import rimraf from './rimraf';
-import { mkdir } from './util';
 import md5 from 'MD5';
 import ncp from 'ncp';
 import B from 'bluebird';
@@ -17,11 +16,20 @@ let fs = {
   },
   exists: function (path) { return this.hasAccess(path); },
   rimraf,
-  mkdir,
+  mkdir: async function mkdir (dirName) {
+    let _mkdir = B.promisify(_fs.mkdir);
+    try {
+      await _mkdir(dirName);
+    } catch (err) {
+      if (err && err.code !== "EEXIST") {
+        throw err;
+      }
+    }
+  },
   copyFile: B.promisify(ncp)
 };
 
-const simples = ['open', 'close', 'mkdir', 'access', 'readFile', 'writeFile',
+const simples = ['open', 'close', 'access', 'readFile', 'writeFile',
                  'write', 'readlink', 'chmod', 'unlink', 'readdir', 'stat', 'rename'];
 
 for (let s of simples) {
