+++ lib/util.js
@@ -1,7 +1,6 @@
 import B from 'bluebird';
 import _ from 'lodash';
 import os from 'os';
-import fs from './fs';
 import path from 'path';
 
 export function hasContent (val) {
@@ -50,16 +49,6 @@ function escapeSpecialChars (str, quoteEscape) {
   return str;
 }
 
-async function mkdir (dirName) {
-  try {
-    await fs.mkdir(dirName);
-  } catch (err) {
-    if (err && err.code !== "EEXIST") {
-      throw err;
-    }
-  }
-}
-
 function localIp () {
   let ip = _.chain(os.networkInterfaces())
     .values()
@@ -97,4 +86,4 @@ function multiResolve (roots, ...args) {
   });
 }
 
-export { hasValue, escapeSpace, escapeSpecialChars, mkdir, localIp, cancellableDelay, multiResolve };
+export { hasValue, escapeSpace, escapeSpecialChars, localIp, cancellableDelay, multiResolve };
